Create Power of Three
O(1) Time Complexity and O(1) Space Complexity. Here, 1162261467 is 3^19 which is the largest power of 3 possible considering 4 bytes size of int in java. Since 3 is a prime number, all its power will also be a prime number and hence will be divisible by only 3^0, 3^1, 3^2, ..... and so on. Just by checking whether 3^19 is divisible by 'n', we can identify if 'n' is the power of three or not.